@@ -1,6 +1,6 @@
 <?php
 
-namespace BinaryCats\LobWebhooks\Tests;
+namespace Tests;
 
 use BinaryCats\LobWebhooks\LobWebhooksServiceProvider;
 use CreateWebhookCallsTable;
@@ -25,21 +25,23 @@ abstract class TestCase extends OrchestraTestCase
      */
     protected function getEnvironmentSetUp($app)
     {
-        $app['config']->set('database.default', 'sqlite');
-        $app['config']->set('database.connections.sqlite', [
+        config()->set('database.default', 'sqlite');
+        config()->set('database.connections.sqlite', [
             'driver'   => 'sqlite',
             'database' => ':memory:',
             'prefix'   => '',
         ]);
-
-        config(['lob-webhooks.signing_secret' => 'secret']);
+        config()->set('lob-webhooks.signing_secret', 'secret');
     }
 
+    /**
+     * @return void
+     */
     protected function setUpDatabase()
     {
-        include_once __DIR__.'/../vendor/spatie/laravel-webhook-client/database/migrations/create_webhook_calls_table.php.stub';
+        $migration = include __DIR__.'/../vendor/spatie/laravel-webhook-client/database/migrations/create_webhook_calls_table.php.stub';
 
-        (new CreateWebhookCallsTable())->up();
+        $migration->up();
     }
 
     /**
